Replace AppAuthenticator with API
